requirements.txt
asgiref==3.6.0
boto3==1.26.133
botocore==1.29.133
Django==4.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-js-asset==2.0.0
django-rest-passwordreset==1.3.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
jdatetime==4.1.1
jmespath==1.0.1
mysqlclient==2.1.1
Pillow==9.5.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.4
typing==3.7.4.3
tzdata==2023.3
urllib3==1.26.15
